move helper functions to formattingHelpers
